+++ src/firecrest/status/health_check/health_checker_cluster.py
@@ -5,7 +5,7 @@
 
 import asyncio
 import time
-
+import logging
 
 from firecrest.config import HPCCluster, HealthCheckException
 
@@ -54,30 +54,38 @@ class ClusterHealthChecker:
             )
             auth = self.token_decoder.auth_from_token(token["access_token"])
             checks = []
-            sechedulerCheck = SchedulerHealthCheck(
-                system=self.cluster,
-                auth=auth,
-                token=token,
-                timeout=self.cluster.probing.timeout,
-            )
-            checks += [sechedulerCheck.check()]
-            sshCheck = SSHHealthCheck(
-                system=self.cluster,
-                auth=auth,
-                token=token,
-                timeout=self.cluster.probing.timeout,
-            )
-            checks += [sshCheck.check()]
+            if self.cluster.probing.services is not None:
+                for service in self.cluster.probing.services:
+                    for k in service.keys():
+                        match k:
+                            case 'scheduler':
+                                check = SchedulerHealthCheck(
+                                    system=self.cluster,
+                                    auth=auth,
+                                    token=token,
+                                    timeout=service[k].timeout,
+                                )
+                            case 'ssh':
+                                check = SSHHealthCheck(
+                                    system=self.cluster,
+                                    auth=auth,
+                                    token=token,
+                                    timeout=service[k].timeout,
+                                )
+                            case 'filesystems':
+                                for filesystem in self.cluster.file_systems:
+                                    check = FilesystemHealthCheck(
+                                        system=self.cluster,
+                                        auth=auth,
+                                        token=token,
+                                        path=filesystem.path,
+                                        timeout=service[k].timeout,
+                                    )
+                            case _:
+                                check = None
 
-            for filesystem in self.cluster.file_systems:
-                filesystemCheck = FilesystemHealthCheck(
-                    system=self.cluster,
-                    auth=auth,
-                    token=token,
-                    path=filesystem.path,
-                    timeout=self.cluster.probing.timeout,
-                )
-                checks += [filesystemCheck.check()]
+                        if check is not None:
+                            checks += [check.check()]
 
             results = await asyncio.gather(*checks, return_exceptions=True)
             self.cluster.servicesHealth = results